refactor: Rename `lib/plugins/aws/utils/findReferences.js`
