+++ lib/plugins/aws/deploy/lib/extendedValidate.js
@@ -2,7 +2,7 @@
 
 const path = require('path');
 const _ = require('lodash');
-const findReferences = require('../../utils/findReferences');
+const findReferences = require('../../utils/find-references');
 const ServerlessError = require('../../../../serverless-error');
 const { log } = require('@serverless/utils/log');
