+++ lib/plugins/aws/package/lib/saveServiceState.js
@@ -2,7 +2,7 @@
 
 const path = require('path');
 const _ = require('lodash');
-const findReferences = require('../../utils/findReferences');
+const findReferences = require('../../utils/find-references');
 
 module.exports = {
   saveServiceState() {
