+++ lib/plugins/aws/utils/find-references.js
